updated the update command
