+++ sdk/src/rest/commands/read/extensions.ts
@@ -1,17 +1,5 @@
+import type { ExtensionItem } from '../../../types/index.js';
 import type { RestCommand } from '../../types.js';
-
-export type ExtensionTypes = "interface" | "display" | "layout" | "module" | "panel" | "hook" | "endpoint" | "operation" | "bundle";
-
-/**
- * Copied from @directus/extensions
- * The API output structure used when engaging with the /extensions endpoints
- */
-export interface ReadExtensionOutput {
-	name: string;
-	bundle: string | null;
-	schema: { type: ExtensionTypes, local: boolean } | null;
-	meta: { enabled: boolean; /* options, permissions? */ };
-}
 
 
 /**
@@ -19,7 +7,7 @@ export interface ReadExtensionOutput {
  * @returns An array of extensions.
  */
 export const readExtensions =
-	<Schema extends object>(): RestCommand<ReadExtensionOutput[], Schema> =>
+	<Schema extends object>(): RestCommand<ExtensionItem[], Schema> =>
 	() => ({
 		path: `/extensions/`,
 		method: 'GET',
