+++ sdk/src/types/index.ts
@@ -1,6 +1,7 @@
 export * from './aggregate.js';
 export * from './client.js';
 export * from './deep.js';
+export * from './extensions.js';
 export * from './fields.js';
 export * from './filters.js';
 export * from './globals.js';
